+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -1,2 +1,2 @@
-41f68d7373b4bba36173f0990869bc49956edd49028d01c05b11b2bf195423fe  plugin-sdk-api-baseline.json
-a3b1536204f91416c84e347e9abb7d66b915225adea19a5bc76457ca413ca99a  plugin-sdk-api-baseline.jsonl
+72a838530460f61e1f8bb76f454cbeb5e4824bebe2db777515fa83a8ba836bb2  plugin-sdk-api-baseline.json
+8f18c66bc06e5c310889e7a8c70b69a74433eef932b6a3d97cf26b10af6c5f3c  plugin-sdk-api-baseline.jsonl
