+++ scripts/plugin-sdk-surface-report.mjs
@@ -208,9 +208,9 @@ export function readPluginSdkSurfaceBudgets(env = process.env) {
     // The focused HTML entity runtime and quote-aware HTML tokenizer add one public function each.
     publicExports: readPluginSdkSurfaceBudgetEnv(
       "OPENCLAW_PLUGIN_SDK_MAX_PUBLIC_EXPORTS",
-      // +3: registerMcpServerConnectionResolver context/result/resolver types (#106229).
+      // +4: registerMcpServerConnectionResolver context/result/resolver/registration types (#106229).
       // +2: materializeRequesterScopedMcpToolsForHarnessRun (agent-harness-runtime + compat mirror).
-      10689,
+      10690,
       env,
     ),
     publicFunctionExports: readPluginSdkSurfaceBudgetEnv(
