+++ src/agents/agent-bundle-mcp-runtime.ts
@@ -1,5 +1,4 @@
-/** Session-scoped MCP runtime manager, catalog loader, and transport lifecycle. */
-import crypto from "node:crypto";
+/** Session-scoped MCP runtime catalog loader and transport lifecycle. */
 import { Client, type ClientOptions } from "@modelcontextprotocol/sdk/client/index.js";
 import { StreamableHTTPClientTransport } from "@modelcontextprotocol/sdk/client/streamableHttp.js";
 import type { Transport } from "@modelcontextprotocol/sdk/shared/transport.js";
@@ -10,18 +9,36 @@ import {
 } from "@modelcontextprotocol/sdk/types.js";
 import type { ServerCapabilities } from "@modelcontextprotocol/sdk/types.js";
 import { redactSensitiveUrlLikeString } from "@openclaw/net-policy/redact-sensitive-url";
-import {
-  normalizeLowercaseStringOrEmpty,
-  normalizeOptionalString,
-} from "@openclaw/normalization-core/string-coerce";
+import { normalizeLowercaseStringOrEmpty } from "@openclaw/normalization-core/string-coerce";
 import type { OpenClawConfig } from "../config/types.openclaw.js";
 import { toErrorObject } from "../infra/errors.js";
 import { logWarn } from "../logger.js";
 import type { PluginManifestRegistry } from "../plugins/manifest-registry.js";
-import { resolveGlobalSingleton } from "../shared/global-singleton.js";
 import { runTasksWithConcurrency } from "../utils/run-with-concurrency.js";
+import { mergeMcpToolCatalogs } from "./agent-bundle-mcp-combined.js";
 import { matchesMcpToolFilterPattern } from "./agent-bundle-mcp-filter.js";
+import {
+  completeDeferredSessionMcpRuntimeRetirement,
+  disposeAllSessionMcpRuntimes,
+  getAdvertisedScopedMcpCatalog,
+  getOrCreateRequesterScopedMcpRuntime,
+  getOrCreateSessionMcpRuntime,
+  getSessionMcpRuntimeManagerForTesting,
+  peekSessionMcpRuntime,
+  rememberAdvertisedScopedMcpCatalog,
+  retireSessionMcpRuntime,
+  retireSessionMcpRuntimeForSessionKey,
+} from "./agent-bundle-mcp-manager-api.js";
+import {
+  createSessionMcpRuntimeManager,
+  setDefaultCreateSessionMcpRuntime,
+} from "./agent-bundle-mcp-manager.js";
 import { assignSafeServerNames, sanitizeServerName } from "./agent-bundle-mcp-names.js";
+import {
+  loadSessionMcpConfig,
+  resolveSessionMcpConfigSummary,
+} from "./agent-bundle-mcp-runtime-config.js";
+import { resolveSessionMcpRuntimeIdleTtlMs } from "./agent-bundle-mcp-runtime-shared.js";
 import type {
   McpCatalogTool,
   McpRequestOptions,
@@ -32,16 +49,9 @@ import type {
   SessionMcpRuntime,
   SessionMcpRuntimeManager,
 } from "./agent-bundle-mcp-types.js";
-import { loadEmbeddedAgentMcpConfig } from "./embedded-agent-mcp.js";
 import { isMcpConfigRecord } from "./mcp-config-shared.js";
 import {
   applyMcpConnectionOverride,
-  buildMcpRequesterRuntimeCacheKey,
-  hashMcpResolvedConnections,
-  partitionMcpServersByConnectionScope,
-  redactMcpServersForFingerprint,
-  resolveMcpConnectionRevalidateMs,
-  resolveRequesterScopedMcpConnections,
   type McpServerConnectionResolved,
 } from "./mcp-connection-resolver.js";
 import { createMcpJsonSchemaValidator } from "./mcp-json-schema-validator.js";
@@ -65,20 +75,9 @@ type BundleMcpSession = {
   detachStderr?: () => void;
 };
 
-type LoadedMcpConfig = ReturnType<typeof loadEmbeddedAgentMcpConfig>;
 type ListedTool = Awaited<ReturnType<Client["listTools"]>>["tools"][number];
-type CreateSessionMcpRuntime = (
-  params: Parameters<typeof createSessionMcpRuntime>[0] & { configFingerprint?: string },
-) => SessionMcpRuntime;
-
-const SESSION_MCP_RUNTIME_MANAGER_KEY = Symbol.for("openclaw.sessionMcpRuntimeManager");
 const MCP_APPS_CLIENT_EXTENSION = "io.modelcontextprotocol/ui";
 const MCP_APP_RESOURCE_MIME_TYPE = "text/html;profile=mcp-app";
-const DEFAULT_SESSION_MCP_RUNTIME_IDLE_TTL_MS = 10 * 60 * 1000;
-const SESSION_MCP_RUNTIME_SWEEP_INTERVAL_MS = 60 * 1000;
-// Bounds live per-sender MCP transports in one session between idle sweeps;
-// far above concurrent-run parallelism, so active requesters never evict.
-const SESSION_MCP_MAX_IDLE_REQUESTER_RUNTIMES = 64;
 const BUNDLE_MCP_FAILURE_THRESHOLD = 3;
 const BUNDLE_MCP_FAILURE_COOLDOWN_MS = 60_000;
 const BUNDLE_MCP_CATALOG_LIST_TIMEOUT_MS = 1_500;
@@ -350,145 +349,8 @@ async function disposeSession(session: BundleMcpSession) {
   }
 }
 
-function digestSafeServerNameAssignments(
-  safeServerNamesByServer?: ReadonlyMap<string, string>,
-): Record<string, string> | undefined {
-  if (!safeServerNamesByServer || safeServerNamesByServer.size === 0) {
-    return undefined;
-  }
-  return Object.fromEntries(
-    [...safeServerNamesByServer.entries()].toSorted(([a], [b]) => a.localeCompare(b)),
-  );
-}
-
-function createCatalogFingerprint(params: {
-  servers: Record<string, unknown>;
-  mcpAppsEnabled: boolean;
-  /** Full-set server→safeName map; assignment changes must invalidate all partitions. */
-  safeServerNames?: Record<string, string>;
-}): string {
-  // Session MCP fingerprints only invalidate in-memory runtime catalogs.
-  // Algorithm changes can cause one cache miss, but no persisted state migration.
-  // Per-user url/headers never enter this hash (see redactMcpServersForFingerprint).
-  return crypto.createHash("sha256").update(JSON.stringify(params)).digest("hex");
-}
-
-function filterMcpServers<T>(
-  mcpServers: Record<string, T>,
-  options?: {
-    includeServerNames?: ReadonlySet<string>;
-    excludeServerNames?: ReadonlySet<string>;
-  },
-): Record<string, T> {
-  if (!options?.includeServerNames && !options?.excludeServerNames) {
-    return mcpServers;
-  }
-  const filtered: Record<string, T> = {};
-  for (const [serverName, rawServer] of Object.entries(mcpServers)) {
-    if (options.includeServerNames && !options.includeServerNames.has(serverName)) {
-      continue;
-    }
-    if (options.excludeServerNames?.has(serverName)) {
-      continue;
-    }
-    filtered[serverName] = rawServer;
-  }
-  return filtered;
-}
-
-function loadSessionMcpConfig(params: {
-  workspaceDir: string;
-  cfg?: OpenClawConfig;
-  logDiagnostics?: boolean;
-  manifestRegistry?: Pick<PluginManifestRegistry, "plugins">;
-  includeServerNames?: ReadonlySet<string>;
-  excludeServerNames?: ReadonlySet<string>;
-  /** Server names whose url/headers must not affect the fingerprint. */
-  redactConnectionServerNames?: ReadonlySet<string>;
-  /** Full-set safe-name assignments; folded into fingerprint for all partitions. */
-  safeServerNamesByServer?: ReadonlyMap<string, string>;
-}): {
-  loaded: LoadedMcpConfig;
-  fingerprint: string;
-} {
-  const loaded = loadEmbeddedAgentMcpConfig({
-    workspaceDir: params.workspaceDir,
-    cfg: params.cfg,
-    manifestRegistry: params.manifestRegistry,
-  });
-  if (params.logDiagnostics !== false) {
-    for (const diagnostic of loaded.diagnostics) {
-      logWarn(`bundle-mcp: ${diagnostic.pluginId}: ${diagnostic.message}`);
-    }
-  }
-  const mcpServers = filterMcpServers(loaded.mcpServers, {
-    includeServerNames: params.includeServerNames,
-    excludeServerNames: params.excludeServerNames,
-  });
-  const fingerprintServers = params.redactConnectionServerNames?.size
-    ? redactMcpServersForFingerprint(mcpServers, params.redactConnectionServerNames)
-    : mcpServers;
-  const safeServerNames = digestSafeServerNameAssignments(params.safeServerNamesByServer);
-  return {
-    loaded: {
-      ...loaded,
-      mcpServers,
-    },
-    fingerprint: createCatalogFingerprint({
-      servers: fingerprintServers,
-      mcpAppsEnabled: params.cfg?.mcp?.apps?.enabled === true,
-      ...(safeServerNames ? { safeServerNames } : {}),
-    }),
-  };
-}
-
-/**
- * Loads enabled MCP config metadata for a session without creating runtimes,
- * connecting transports, or issuing MCP tools/list requests.
- */
-export function resolveSessionMcpConfigSummary(params: {
-  workspaceDir: string;
-  cfg?: OpenClawConfig;
-  manifestRegistry?: Pick<PluginManifestRegistry, "plugins">;
-}): { fingerprint: string; serverNames: string[] } {
-  const { loaded, fingerprint } = loadSessionMcpConfig({
-    workspaceDir: params.workspaceDir,
-    cfg: params.cfg,
-    logDiagnostics: false,
-    manifestRegistry: params.manifestRegistry,
-  });
-  const serverNames = Object.keys(loaded.mcpServers).toSorted((a, b) => a.localeCompare(b));
-  if (serverNames.length === 0) {
-    return { fingerprint, serverNames };
-  }
-  // Mirror getOrCreate: the bare-keyed runtime folds full-set safe names into
-  // its fingerprint and excludes requester-scoped servers from its partition.
-  // Compare apples-to-apples or tools.effective reports stale-config forever.
-  const safeServerNamesByServer = assignSafeServerNames(Object.keys(loaded.mcpServers));
-  const { requesterScopedServerNames } = partitionMcpServersByConnectionScope(loaded.mcpServers);
-  const { fingerprint: bareRuntimeFingerprint } = loadSessionMcpConfig({
-    workspaceDir: params.workspaceDir,
-    cfg: params.cfg,
-    logDiagnostics: false,
-    manifestRegistry: params.manifestRegistry,
-    ...(requesterScopedServerNames.length > 0
-      ? { excludeServerNames: new Set(requesterScopedServerNames) }
-      : {}),
-    safeServerNamesByServer,
-  });
-  return { fingerprint: bareRuntimeFingerprint, serverNames };
-}
-
 function createDisposedError(sessionId: string): Error {
   return new Error(`bundle-mcp runtime disposed for session ${sessionId}`);
-}
-
-function resolveSessionMcpRuntimeIdleTtlMs(cfg?: OpenClawConfig): number {
-  const raw = cfg?.mcp?.sessionIdleTtlMs;
-  if (typeof raw === "number" && Number.isFinite(raw) && raw >= 0) {
-    return Math.floor(raw);
-  }
-  return DEFAULT_SESSION_MCP_RUNTIME_IDLE_TTL_MS;
 }
 
 export function createSessionMcpRuntime(params: {
@@ -1086,1290 +948,22 @@ export function createSessionMcpRuntime(params: {
   };
 }
 
-const COMBINED_SESSION_MCP_RUNTIME = Symbol.for("openclaw.combinedSessionMcpRuntime");
+setDefaultCreateSessionMcpRuntime(createSessionMcpRuntime);
 
-type CombinedSessionMcpRuntime = SessionMcpRuntime & {
-  [COMBINED_SESSION_MCP_RUNTIME]: true;
-  managedParts: readonly SessionMcpRuntime[];
+export {
+  completeDeferredSessionMcpRuntimeRetirement,
+  disposeAllSessionMcpRuntimes,
+  getAdvertisedScopedMcpCatalog,
+  getOrCreateRequesterScopedMcpRuntime,
+  getOrCreateSessionMcpRuntime,
+  peekSessionMcpRuntime,
+  rememberAdvertisedScopedMcpCatalog,
+  resolveSessionMcpConfigSummary,
+  retireSessionMcpRuntime,
+  retireSessionMcpRuntimeForSessionKey,
 };
-
-function isCombinedSessionMcpRuntime(
-  runtime: SessionMcpRuntime,
-): runtime is CombinedSessionMcpRuntime {
-  return (runtime as CombinedSessionMcpRuntime)[COMBINED_SESSION_MCP_RUNTIME] !== undefined;
-}
-
-function parseRuntimeCacheSessionId(runtimeKey: string): string {
-  if (!runtimeKey.startsWith("{")) {
-    return runtimeKey;
-  }
-  try {
-    const parsed = JSON.parse(runtimeKey) as { sessionId?: unknown };
-    return typeof parsed.sessionId === "string" ? parsed.sessionId : runtimeKey;
-  } catch {
-    return runtimeKey;
-  }
-}
-
-/**
- * Merge catalogs from static + requester partitions.
- * Safe names are precomputed from the full declared set, so no re-suffix is needed.
- */
-function mergeMcpToolCatalogs(catalogs: readonly McpToolCatalog[]): McpToolCatalog {
-  const servers: Record<string, McpServerCatalog> = {};
-  const tools: McpCatalogTool[] = [];
-  const diagnostics: McpToolCatalogDiagnostic[] = [];
-
-  for (const catalog of catalogs) {
-    for (const [serverName, server] of Object.entries(catalog.servers).toSorted(([a], [b]) =>
-      a.localeCompare(b),
-    )) {
-      servers[serverName] = server;
-    }
-    tools.push(...catalog.tools);
-    if (catalog.diagnostics) {
-      diagnostics.push(...catalog.diagnostics);
-    }
-  }
-  tools.sort((a, b) => {
-    const serverOrder = a.safeServerName.localeCompare(b.safeServerName);
-    if (serverOrder !== 0) {
-      return serverOrder;
-    }
-    const toolOrder = a.toolName.localeCompare(b.toolName);
-    if (toolOrder !== 0) {
-      return toolOrder;
-    }
-    return a.serverName.localeCompare(b.serverName);
-  });
-  return {
-    version: 1,
-    generatedAt: Math.max(0, ...catalogs.map((catalog) => catalog.generatedAt)),
-    servers,
-    tools,
-    ...(diagnostics.length > 0 ? { diagnostics } : {}),
-  };
-}
-
-function createCombinedSessionMcpRuntime(params: {
-  sessionId: string;
-  sessionKey?: string;
-  workspaceDir: string;
-  agentDir?: string;
-  parts: readonly SessionMcpRuntime[];
-}): SessionMcpRuntime {
-  if (params.parts.length === 1) {
-    return params.parts[0]!;
-  }
-  const parts = params.parts;
-  let lastUsedAt = Math.max(...parts.map((part) => part.lastUsedAt));
-  let cachedCatalog: McpToolCatalog | null = null;
-  let mergedSourceCatalogs: ReadonlyArray<McpToolCatalog> | null = null;
-  let catalogInFlight: Promise<McpToolCatalog> | undefined;
-  const serverOwner = new Map<string, SessionMcpRuntime>();
-
-  const rememberServerOwners = (catalog: McpToolCatalog, owner: SessionMcpRuntime) => {
-    for (const serverName of Object.keys(catalog.servers)) {
-      serverOwner.set(serverName, owner);
-    }
-  };
-
-  // Parts invalidate their own catalogs on tools/list_changed by replacing or
-  // clearing the cached object. Identity-compare against what was merged so the
-  // facade re-merges instead of serving a stale combined catalog.
-  const cachedCatalogIsCurrent = (): boolean =>
-    cachedCatalog !== null &&
-    mergedSourceCatalogs !== null &&
-    parts.every((part, index) => part.peekCatalog() === mergedSourceCatalogs?.[index]);
-
-  const loadCatalog = async (): Promise<McpToolCatalog> => {
-    if (cachedCatalog && cachedCatalogIsCurrent()) {
-      return cachedCatalog;
-    }
-    if (catalogInFlight) {
-      return catalogInFlight;
-    }
-    const inFlight = (async () => {
-      const catalogs = await Promise.all(parts.map((part) => part.getCatalog()));
-      serverOwner.clear();
-      for (let index = 0; index < parts.length; index += 1) {
-        rememberServerOwners(catalogs[index]!, parts[index]!);
-      }
-      mergedSourceCatalogs = catalogs;
-      cachedCatalog = mergeMcpToolCatalogs(catalogs);
-      return cachedCatalog;
-    })();
-    catalogInFlight = inFlight;
-    try {
-      return await inFlight;
-    } finally {
-      if (catalogInFlight === inFlight) {
-        catalogInFlight = undefined;
-      }
-    }
-  };
-
-  // Fresh combined facades have an empty owner map until the catalog is loaded.
-  // Share one in-flight getCatalog so concurrent tool/resource calls do not fan out.
-  const ownerForServer = async (serverName: string): Promise<SessionMcpRuntime> => {
-    if (serverOwner.size === 0) {
-      await loadCatalog();
-    }
-    const owner = serverOwner.get(serverName);
-    if (owner) {
-      return owner;
-    }
-    throw new Error(`bundle-mcp server "${serverName}" is not connected`);
-  };
-
-  const combined: CombinedSessionMcpRuntime = {
-    [COMBINED_SESSION_MCP_RUNTIME]: true,
-    managedParts: parts,
-    sessionId: params.sessionId,
-    sessionKey: params.sessionKey,
-    workspaceDir: params.workspaceDir,
-    agentDir: params.agentDir,
-    configFingerprint: parts.map((part) => part.configFingerprint).join(":"),
-    isRequesterScopedServer(serverName) {
-      // Owner map is populated by the catalog load that exposed the tool.
-      return serverOwner.get(serverName)?.requesterScope !== undefined;
-    },
-    mcpAppsEnabled: parts.some((part) => part.mcpAppsEnabled === true),
-    createdAt: Math.min(...parts.map((part) => part.createdAt)),
-    get lastUsedAt() {
-      return lastUsedAt;
-    },
-    get activeLeases() {
-      return parts.reduce((sum, part) => sum + (part.activeLeases ?? 0), 0);
-    },
-    acquireLease() {
-      const releases = parts.map((part) => part.acquireLease?.());
-      let released = false;
-      return () => {
-        if (released) {
-          return;
-        }
-        released = true;
-        for (const release of releases) {
-          release?.();
-        }
-      };
-    },
-    getCatalog: loadCatalog,
-    peekCatalog() {
-      if (cachedCatalog && cachedCatalogIsCurrent()) {
-        return cachedCatalog;
-      }
-      const peeked = parts.map((part) => part.peekCatalog());
-      if (peeked.some((catalog) => catalog === null)) {
-        return null;
-      }
-      return mergeMcpToolCatalogs(peeked as McpToolCatalog[]);
-    },
-    markUsed() {
-      lastUsedAt = Date.now();
-      for (const part of parts) {
-        part.markUsed();
-      }
-    },
-    async callTool(serverName, toolName, input) {
-      return await (await ownerForServer(serverName)).callTool(serverName, toolName, input);
-    },
-    async listTools(serverName, requestParams) {
-      const owner = await ownerForServer(serverName);
-      if (!owner.listTools) {
-        throw new Error(`bundle-mcp server "${serverName}" does not support listTools`);
-      }
-      return await owner.listTools(serverName, requestParams);
-    },
-    async listResources(serverName, options) {
-      const owner = await ownerForServer(serverName);
-      if (!owner.listResources) {
-        throw new Error(`bundle-mcp server "${serverName}" does not support listResources`);
-      }
-      return await owner.listResources(serverName, options);
-    },
-    async readResource(serverName, uri, options) {
-      const owner = await ownerForServer(serverName);
-      if (!owner.readResource) {
-        throw new Error(`bundle-mcp server "${serverName}" does not support readResource`);
-      }
-      return await owner.readResource(serverName, uri, options);
-    },
-    async listResourceTemplates(serverName, requestParams) {
-      const owner = await ownerForServer(serverName);
-      if (!owner.listResourceTemplates) {
-        throw new Error(`bundle-mcp server "${serverName}" does not support listResourceTemplates`);
-      }
-      return await owner.listResourceTemplates(serverName, requestParams);
-    },
-    async listPrompts(serverName) {
-      const owner = await ownerForServer(serverName);
-      if (!owner.listPrompts) {
-        throw new Error(`bundle-mcp server "${serverName}" does not support listPrompts`);
-      }
-      return await owner.listPrompts(serverName);
-    },
-    async getPrompt(serverName, name, args) {
-      const owner = await ownerForServer(serverName);
-      if (!owner.getPrompt) {
-        throw new Error(`bundle-mcp server "${serverName}" does not support getPrompt`);
-      }
-      return await owner.getPrompt(serverName, name, args);
-    },
-    async dispose() {
-      await Promise.allSettled(parts.map((part) => part.dispose()));
-    },
-  };
-  return combined;
-}
-
-function createSessionMcpRuntimeManager(
-  opts: {
-    createRuntime?: CreateSessionMcpRuntime;
-    now?: () => number;
-    enableIdleSweepTimer?: boolean;
-    idleSweepIntervalMs?: number;
-    maxIdleRequesterRuntimesPerSession?: number;
-  } = {},
-): SessionMcpRuntimeManager {
-  // Keys are bare sessionId for static runtimes, or requester composite JSON keys.
-  const runtimesBySessionId = new Map<string, SessionMcpRuntime>();
-  const sessionIdBySessionKey = new Map<string, string>();
-  const idleTtlMsBySessionId = new Map<string, number>();
-  const deferredRetirementSessionIds = new Set<string>();
-  // Manager-side only: connection hash + resolve time. Never stores raw url/headers.
-  const connectionMetaByRuntimeKey = new Map<
-    string,
-    { connectionHash: string; resolvedAt: number }
-  >();
-  /**
-   * Session-stable advertised catalogs for requester-scoped servers.
-   * Keyed by sessionId → serverName. Specs must not vary per sender or shared
-   * Codex threads rotate (dynamicToolsFingerprint churn).
-   */
-  const advertisedScopedCatalogBySessionId = new Map<
-    string,
-    {
-      servers: Map<string, McpServerCatalog>;
-      toolsByServer: Map<string, McpCatalogTool[]>;
-      signaturesByServer: Map<string, string>;
-    }
-  >();
-  /**
-   * Per-runtimeKey serialization for requester resolve+install and dispose.
-   * Sections never overlap for one key, so a slow resolve cannot clobber a newer install.
-   * Entries are removed when their chain drains.
-   */
-  const requesterWorkChains = new Map<string, Promise<unknown>>();
-  const createRuntime = opts.createRuntime ?? createSessionMcpRuntime;
-  const now = opts.now ?? Date.now;
-  // Static bare-sessionId create dedup only. Requester keys use requesterWorkChains exclusively.
-  const createInFlight = new Map<
-    string,
-    {
-      promise: Promise<SessionMcpRuntime>;
-      workspaceDir: string;
-      agentDir?: string;
-      configFingerprint: string;
-    }
-  >();
-  const idleSweepIntervalMs = opts.idleSweepIntervalMs ?? SESSION_MCP_RUNTIME_SWEEP_INTERVAL_MS;
-  let idleSweepTimer: ReturnType<typeof setInterval> | undefined;
-  let idleSweepInFlight: Promise<void> | undefined;
-
-  const forgetSessionKeysForSessionId = (sessionId: string) => {
-    for (const [sessionKey, mappedSessionId] of sessionIdBySessionKey.entries()) {
-      if (mappedSessionId === sessionId) {
-        sessionIdBySessionKey.delete(sessionKey);
-      }
-    }
-  };
-
-  const runtimeKeysForSessionId = (sessionId: string): string[] => {
-    const keys: string[] = [];
-    for (const [runtimeKey, runtime] of runtimesBySessionId.entries()) {
-      if (runtime.sessionId === sessionId) {
-        keys.push(runtimeKey);
-      }
-    }
-    return keys;
-  };
-
-  const totalActiveLeasesForSessionId = (sessionId: string): number => {
-    let total = 0;
-    for (const runtimeKey of runtimeKeysForSessionId(sessionId)) {
-      total += runtimesBySessionId.get(runtimeKey)?.activeLeases ?? 0;
-    }
-    return total;
-  };
-
-  const runExclusiveOnRuntimeKey = <T>(runtimeKey: string, work: () => Promise<T>): Promise<T> => {
-    const previous = requesterWorkChains.get(runtimeKey) ?? Promise.resolve();
-    const run = previous.catch(() => undefined).then(() => work());
-    const settled: Promise<unknown> = run.then(
-      () => undefined,
-      () => undefined,
-    );
-    requesterWorkChains.set(runtimeKey, settled);
-    void settled.finally(() => {
-      if (requesterWorkChains.get(runtimeKey) === settled) {
-        requesterWorkChains.delete(runtimeKey);
-      }
-    });
-    return run;
-  };
-
-  const sweepIdleRuntimes = async (): Promise<number> => {
-    const nowMs = now();
-    const expired: SessionMcpRuntime[] = [];
-    for (const [runtimeKey, runtime] of runtimesBySessionId.entries()) {
-      const idleTtlMs =
-        idleTtlMsBySessionId.get(runtimeKey) ??
-        idleTtlMsBySessionId.get(runtime.sessionId) ??
-        DEFAULT_SESSION_MCP_RUNTIME_IDLE_TTL_MS;
-      if (idleTtlMs <= 0 || (runtime.activeLeases ?? 0) > 0) {
-        continue;
-      }
-      if (nowMs - runtime.lastUsedAt < idleTtlMs) {
-        continue;
-      }
-      runtimesBySessionId.delete(runtimeKey);
-      idleTtlMsBySessionId.delete(runtimeKey);
-      connectionMetaByRuntimeKey.delete(runtimeKey);
-      expired.push(runtime);
-    }
-    const touchedSessionIds = new Set(expired.map((runtime) => runtime.sessionId));
-    for (const sessionId of touchedSessionIds) {
-      if (runtimeKeysForSessionId(sessionId).length === 0) {
-        deferredRetirementSessionIds.delete(sessionId);
-        forgetSessionKeysForSessionId(sessionId);
-      }
-    }
-    await Promise.allSettled(expired.map((runtime) => runtime.dispose()));
-    return expired.length;
-  };
-
-  const maxIdleRequesterRuntimes =
-    opts.maxIdleRequesterRuntimesPerSession ?? SESSION_MCP_MAX_IDLE_REQUESTER_RUNTIMES;
-
-  /**
-   * A busy shared channel can otherwise accumulate one live scoped runtime per
-   * sender until the idle TTL fires. Evict LRU zero-lease requester runtimes
-   * beyond the cap; leased runtimes and the bare static runtime never evict.
-   */
-  const enforceRequesterRuntimeCap = async (
-    sessionId: string,
-    keepRuntimeKey: string,
-  ): Promise<void> => {
-    const requesterKeys = runtimeKeysForSessionId(sessionId).filter(
-      (runtimeKey) => runtimeKey !== sessionId,
-    );
-    const overflow = requesterKeys.length - maxIdleRequesterRuntimes;
-    if (overflow <= 0) {
-      return;
-    }
-    const evictable = requesterKeys
-      .filter((runtimeKey) => runtimeKey !== keepRuntimeKey)
-      .map((runtimeKey) => ({ runtimeKey, runtime: runtimesBySessionId.get(runtimeKey) }))
-      .filter(
-        (entry): entry is { runtimeKey: string; runtime: SessionMcpRuntime } =>
-          entry.runtime !== undefined && (entry.runtime.activeLeases ?? 0) === 0,
-      )
-      .toSorted((a, b) => a.runtime.lastUsedAt - b.runtime.lastUsedAt)
-      .slice(0, overflow);
-    for (const { runtimeKey, runtime } of evictable) {
-      // Serialize with in-flight work on that key so eviction cannot clobber a
-      // concurrent reuse or install for the same requester.
-      await runExclusiveOnRuntimeKey(runtimeKey, async () => {
-        const current = runtimesBySessionId.get(runtimeKey);
-        if (current !== runtime || (current.activeLeases ?? 0) > 0) {
-          return;
-        }
-        runtimesBySessionId.delete(runtimeKey);
-        idleTtlMsBySessionId.delete(runtimeKey);
-        connectionMetaByRuntimeKey.delete(runtimeKey);
-        await current.dispose();
-      });
-    }
-  };
-
-  const queueIdleSweep = () => {
-    if (idleSweepInFlight) {
-      return;
-    }
-    idleSweepInFlight = sweepIdleRuntimes()
-      .then(() => undefined)
-      .catch((error: unknown) => {
-        logWarn(`bundle-mcp: idle runtime sweep failed: ${String(error)}`);
-      })
-      .finally(() => {
-        idleSweepInFlight = undefined;
-      });
-  };
-
-  const ensureIdleSweepTimer = () => {
-    if (opts.enableIdleSweepTimer === false || idleSweepIntervalMs <= 0 || idleSweepTimer) {
-      return;
-    }
-    idleSweepTimer = setInterval(queueIdleSweep, idleSweepIntervalMs);
-    idleSweepTimer.unref?.();
-  };
-
-  const clearIdleSweepTimer = () => {
-    if (!idleSweepTimer) {
-      return;
-    }
-    clearInterval(idleSweepTimer);
-    idleSweepTimer = undefined;
-  };
-
-  const disposeRuntimeKeyNow = async (runtimeKey: string): Promise<void> => {
-    const inFlight = createInFlight.get(runtimeKey);
-    createInFlight.delete(runtimeKey);
-    let runtime = runtimesBySessionId.get(runtimeKey);
-    if (!runtime && inFlight) {
-      runtime = await inFlight.promise.catch(() => undefined);
-    }
-    runtimesBySessionId.delete(runtimeKey);
-    idleTtlMsBySessionId.delete(runtimeKey);
-    connectionMetaByRuntimeKey.delete(runtimeKey);
-    if (runtime) {
-      await runtime.dispose();
-    }
-  };
-
-  const disposeManagedSession = async (sessionId: string): Promise<void> => {
-    deferredRetirementSessionIds.delete(sessionId);
-    advertisedScopedCatalogBySessionId.delete(sessionId);
-    const runtimeKeys = new Set(runtimeKeysForSessionId(sessionId));
-    for (const runtimeKey of createInFlight.keys()) {
-      if (parseRuntimeCacheSessionId(runtimeKey) === sessionId) {
-        runtimeKeys.add(runtimeKey);
-      }
-    }
-    for (const runtimeKey of requesterWorkChains.keys()) {
-      if (parseRuntimeCacheSessionId(runtimeKey) === sessionId) {
-        runtimeKeys.add(runtimeKey);
-      }
-    }
-    // Serialize disposal with in-flight requester work for composite keys.
-    await Promise.allSettled(
-      [...runtimeKeys].map((runtimeKey) =>
-        runtimeKey.startsWith("{")
-          ? runExclusiveOnRuntimeKey(runtimeKey, () => disposeRuntimeKeyNow(runtimeKey))
-          : disposeRuntimeKeyNow(runtimeKey),
-      ),
-    );
-    forgetSessionKeysForSessionId(sessionId);
-  };
-
-  function scopedCatalogToolsSignature(tools: readonly McpCatalogTool[]): string {
-    return JSON.stringify(
-      tools.map((tool) => [
-        tool.serverName,
-        tool.safeServerName,
-        tool.toolName,
-        tool.title ?? "",
-        tool.description ?? "",
-        tool.fallbackDescription,
-        tool.inputSchema,
-        tool.uiResourceUri ?? "",
-        tool.uiVisibility ?? null,
-      ]),
-    );
-  }
-
-  const rememberAdvertisedScopedCatalog = (sessionId: string, catalog: McpToolCatalog): void => {
-    let entry = advertisedScopedCatalogBySessionId.get(sessionId);
-    if (!entry) {
-      entry = {
-        servers: new Map(),
-        toolsByServer: new Map(),
-        signaturesByServer: new Map(),
-      };
-      advertisedScopedCatalogBySessionId.set(sessionId, entry);
-    }
-    const toolsByServerName = new Map<string, McpCatalogTool[]>();
-    for (const tool of catalog.tools) {
-      const list = toolsByServerName.get(tool.serverName) ?? [];
-      list.push(tool);
-      toolsByServerName.set(tool.serverName, list);
-    }
-    for (const [serverName, server] of Object.entries(catalog.servers)) {
-      const tools = (toolsByServerName.get(serverName) ?? []).toSorted((a, b) =>
-        a.toolName.localeCompare(b.toolName),
-      );
-      const signature = scopedCatalogToolsSignature(tools);
-      // Identity compare: overwrite only when the listed tool surface changes.
-      if (entry.signaturesByServer.get(serverName) === signature) {
-        continue;
-      }
-      entry.servers.set(serverName, server);
-      entry.toolsByServer.set(serverName, tools);
-      entry.signaturesByServer.set(serverName, signature);
-    }
-  };
-
-  const getAdvertisedScopedCatalog = (sessionId: string): McpToolCatalog | null => {
-    const entry = advertisedScopedCatalogBySessionId.get(sessionId);
-    if (!entry || entry.servers.size === 0) {
-      return null;
-    }
-    const servers: Record<string, McpServerCatalog> = {};
-    const tools: McpCatalogTool[] = [];
-    for (const serverName of [...entry.servers.keys()].toSorted((a, b) => a.localeCompare(b))) {
-      servers[serverName] = entry.servers.get(serverName)!;
-      tools.push(...(entry.toolsByServer.get(serverName) ?? []));
-    }
-    tools.sort((a, b) => {
-      const serverOrder = a.safeServerName.localeCompare(b.safeServerName);
-      if (serverOrder !== 0) {
-        return serverOrder;
-      }
-      return a.toolName.localeCompare(b.toolName);
-    });
-    return {
-      version: 1,
-      generatedAt: now(),
-      servers,
-      tools,
-    };
-  };
-
-  type RuntimeEntryParams = {
-    runtimeKey: string;
-    sessionId: string;
-    sessionKey?: string;
-    workspaceDir: string;
-    agentDir?: string;
-    cfg?: OpenClawConfig;
-    manifestRegistry?: Pick<PluginManifestRegistry, "plugins">;
-    idleTtlMs: number;
-    includeServerNames?: ReadonlySet<string>;
-    excludeServerNames?: ReadonlySet<string>;
-    safeServerNamesByServer?: ReadonlyMap<string, string>;
-    connectionOverrides?: ReadonlyMap<string, McpServerConnectionResolved>;
-    redactConnectionServerNames?: ReadonlySet<string>;
-    requesterScope?: SessionMcpRequesterScope;
-    configFingerprint?: string;
-  };
-
-  const matchesStaticReuse = (params: {
-    workspaceDir: string;
-    agentDir?: string;
-    configFingerprint: string;
-    candidate: { workspaceDir: string; agentDir?: string; configFingerprint: string };
-  }): boolean =>
-    params.candidate.workspaceDir === params.workspaceDir &&
-    params.candidate.agentDir === params.agentDir &&
-    params.candidate.configFingerprint === params.configFingerprint;
-
-  /** Static/session runtime get-or-create (createInFlight dedup for bare keys only). */
-  const getOrCreateRuntimeEntry = async (
-    params: RuntimeEntryParams,
-  ): Promise<SessionMcpRuntime> => {
-    const nextFingerprint =
-      params.configFingerprint ??
-      loadSessionMcpConfig({
-        workspaceDir: params.workspaceDir,
-        cfg: params.cfg,
-        logDiagnostics: false,
-        manifestRegistry: params.manifestRegistry,
-        includeServerNames: params.includeServerNames,
-        excludeServerNames: params.excludeServerNames,
-        redactConnectionServerNames: params.redactConnectionServerNames,
-        safeServerNamesByServer: params.safeServerNamesByServer,
-      }).fingerprint;
-    const existing = runtimesBySessionId.get(params.runtimeKey);
-    if (existing) {
-      if (
-        !matchesStaticReuse({
-          workspaceDir: params.workspaceDir,
-          agentDir: params.agentDir,
-          configFingerprint: nextFingerprint,
-          candidate: existing,
-        })
-      ) {
-        runtimesBySessionId.delete(params.runtimeKey);
-        idleTtlMsBySessionId.delete(params.runtimeKey);
-        connectionMetaByRuntimeKey.delete(params.runtimeKey);
-        await existing.dispose();
-      } else {
-        deferredRetirementSessionIds.delete(params.sessionId);
-        existing.markUsed();
-        idleTtlMsBySessionId.set(params.runtimeKey, params.idleTtlMs);
-        return existing;
-      }
-    }
-    const inFlight = createInFlight.get(params.runtimeKey);
-    if (inFlight) {
-      if (
-        matchesStaticReuse({
-          workspaceDir: params.workspaceDir,
-          agentDir: params.agentDir,
-          configFingerprint: nextFingerprint,
-          candidate: inFlight,
-        })
-      ) {
-        return inFlight.promise;
-      }
-      createInFlight.delete(params.runtimeKey);
-      const staleRuntime = await inFlight.promise.catch(() => undefined);
-      runtimesBySessionId.delete(params.runtimeKey);
-      idleTtlMsBySessionId.delete(params.runtimeKey);
-      connectionMetaByRuntimeKey.delete(params.runtimeKey);
-      await staleRuntime?.dispose();
-    }
-    const created = Promise.resolve(
-      createRuntime({
-        sessionId: params.sessionId,
-        sessionKey: params.sessionKey,
-        workspaceDir: params.workspaceDir,
-        agentDir: params.agentDir,
-        cfg: params.cfg,
-        manifestRegistry: params.manifestRegistry,
-        includeServerNames: params.includeServerNames,
-        excludeServerNames: params.excludeServerNames,
-        safeServerNamesByServer: params.safeServerNamesByServer,
-        connectionOverrides: params.connectionOverrides,
-        redactConnectionServerNames: params.redactConnectionServerNames,
-        requesterScope: params.requesterScope,
-        configFingerprint: nextFingerprint,
-      }),
-    ).then((runtime) => {
-      deferredRetirementSessionIds.delete(params.sessionId);
-      runtime.markUsed();
-      runtimesBySessionId.set(params.runtimeKey, runtime);
-      idleTtlMsBySessionId.set(params.runtimeKey, params.idleTtlMs);
-      return runtime;
-    });
-    createInFlight.set(params.runtimeKey, {
-      promise: created,
-      workspaceDir: params.workspaceDir,
-      agentDir: params.agentDir,
-      configFingerprint: nextFingerprint,
-    });
-    try {
-      return await created;
-    } finally {
-      createInFlight.delete(params.runtimeKey);
-    }
-  };
-
-  /**
-   * Install or reuse a requester runtime for already-resolved connections.
-   * Must run inside runExclusiveOnRuntimeKey for this runtimeKey.
-   */
-  const installRequesterRuntime = async (params: {
-    runtimeKey: string;
-    sessionId: string;
-    sessionKey?: string;
-    workspaceDir: string;
-    agentDir?: string;
-    cfg?: OpenClawConfig;
-    manifestRegistry?: Pick<PluginManifestRegistry, "plugins">;
-    idleTtlMs: number;
-    safeServerNamesByServer: ReadonlyMap<string, string>;
-    connectionOverrides: Map<string, McpServerConnectionResolved>;
-    redactConnectionServerNames: ReadonlySet<string>;
-    requesterScope: SessionMcpRequesterScope;
-  }): Promise<SessionMcpRuntime> => {
-    const resolvedNameSet = new Set(params.connectionOverrides.keys());
-    const { fingerprint: resolvedFingerprint } = loadSessionMcpConfig({
-      workspaceDir: params.workspaceDir,
-      cfg: params.cfg,
-      logDiagnostics: false,
-      manifestRegistry: params.manifestRegistry,
-      includeServerNames: resolvedNameSet,
-      redactConnectionServerNames: params.redactConnectionServerNames,
-      safeServerNamesByServer: params.safeServerNamesByServer,
-    });
-    const connectionHash = hashMcpResolvedConnections(params.connectionOverrides);
-    const existing = runtimesBySessionId.get(params.runtimeKey);
-    const meta = connectionMetaByRuntimeKey.get(params.runtimeKey);
-    if (
-      existing &&
-      meta?.connectionHash === connectionHash &&
-      matchesStaticReuse({
-        workspaceDir: params.workspaceDir,
-        agentDir: params.agentDir,
-        configFingerprint: resolvedFingerprint,
-        candidate: existing,
-      })
-    ) {
-      deferredRetirementSessionIds.delete(params.sessionId);
-      existing.markUsed();
-      idleTtlMsBySessionId.set(params.runtimeKey, params.idleTtlMs);
-      connectionMetaByRuntimeKey.set(params.runtimeKey, {
-        connectionHash,
-        resolvedAt: now(),
-      });
-      return existing;
-    }
-    if (existing) {
-      runtimesBySessionId.delete(params.runtimeKey);
-      idleTtlMsBySessionId.delete(params.runtimeKey);
-      connectionMetaByRuntimeKey.delete(params.runtimeKey);
-      await existing.dispose();
-    }
-    const runtime = await getOrCreateRuntimeEntry({
-      runtimeKey: params.runtimeKey,
-      sessionId: params.sessionId,
-      sessionKey: params.sessionKey,
-      workspaceDir: params.workspaceDir,
-      agentDir: params.agentDir,
-      cfg: params.cfg,
-      manifestRegistry: params.manifestRegistry,
-      idleTtlMs: params.idleTtlMs,
-      includeServerNames: resolvedNameSet,
-      safeServerNamesByServer: params.safeServerNamesByServer,
-      connectionOverrides: params.connectionOverrides,
-      redactConnectionServerNames: params.redactConnectionServerNames,
-      requesterScope: params.requesterScope,
-      configFingerprint: resolvedFingerprint,
-    });
-    connectionMetaByRuntimeKey.set(params.runtimeKey, {
-      connectionHash,
-      resolvedAt: now(),
-    });
-    return runtime;
-  };
-
-  /** Revoke cached scoped runtime (empty re-resolution). Auth boundary: leases do not block. */
-  const revokeRequesterRuntime = async (runtimeKey: string): Promise<void> => {
-    await disposeRuntimeKeyNow(runtimeKey);
-  };
-
-  /**
-   * Full requester section for one runtimeKey: reuse / resolve / install / revoke.
-   * Always invoked under runExclusiveOnRuntimeKey.
-   */
-  const resolveAndInstallRequesterRuntime = async (params: {
-    runtimeKey: string;
-    sessionId: string;
-    sessionKey?: string;
-    workspaceDir: string;
-    agentDir?: string;
-    cfg?: OpenClawConfig;
-    manifestRegistry?: Pick<PluginManifestRegistry, "plugins">;
-    idleTtlMs: number;
-    requesterScopedServerNames: readonly string[];
-    scopedNameSet: ReadonlySet<string>;
-    safeServerNamesByServer: ReadonlyMap<string, string>;
-    fullScopedFingerprint: string;
-    requesterSenderId: string;
-    agentAccountId?: string | null;
-    messageChannel?: string | null;
-    requesterScope: SessionMcpRequesterScope;
-  }): Promise<SessionMcpRuntime | undefined> => {
-    const existing = runtimesBySessionId.get(params.runtimeKey);
-    const meta = connectionMetaByRuntimeKey.get(params.runtimeKey);
-    const revalidateMs = resolveMcpConnectionRevalidateMs();
-    // Full-set + within revalidation window: skip resolver I/O.
-    // Revocation/rotation takes effect within MCP_CONNECTION_REVALIDATE_MS even for
-    // continuously active requesters (markUsed does not extend this clock alone).
-    const withinRevalidateWindow = meta !== undefined && now() - meta.resolvedAt < revalidateMs;
-    if (
-      withinRevalidateWindow &&
-      existing &&
-      matchesStaticReuse({
-        workspaceDir: params.workspaceDir,
-        agentDir: params.agentDir,
-        configFingerprint: params.fullScopedFingerprint,
-        candidate: existing,
-      })
-    ) {
-      deferredRetirementSessionIds.delete(params.sessionId);
-      existing.markUsed();
-      idleTtlMsBySessionId.set(params.runtimeKey, params.idleTtlMs);
-      return existing;
-    }
-
-    const connectionOverrides = await resolveRequesterScopedMcpConnections({
-      serverNames: params.requesterScopedServerNames,
-      requesterSenderId: params.requesterSenderId,
-      agentAccountId: params.agentAccountId,
-      messageChannel: params.messageChannel,
-    });
-    if (connectionOverrides.size === 0) {
-      // Empty re-resolution revokes cached scoped credentials.
-      // Leases do not block: this is an authorization boundary.
-      if (runtimesBySessionId.has(params.runtimeKey) || createInFlight.has(params.runtimeKey)) {
-        await revokeRequesterRuntime(params.runtimeKey);
-      }
-      return undefined;
-    }
-    return await installRequesterRuntime({
-      runtimeKey: params.runtimeKey,
-      sessionId: params.sessionId,
-      sessionKey: params.sessionKey,
-      workspaceDir: params.workspaceDir,
-      agentDir: params.agentDir,
-      cfg: params.cfg,
-      manifestRegistry: params.manifestRegistry,
-      idleTtlMs: params.idleTtlMs,
-      safeServerNamesByServer: params.safeServerNamesByServer,
-      connectionOverrides,
-      redactConnectionServerNames: params.scopedNameSet,
-      requesterScope: params.requesterScope,
-    });
-  };
-
-  const manager: SessionMcpRuntimeManager = {
-    async getOrCreate(params) {
-      const idleTtlMs = resolveSessionMcpRuntimeIdleTtlMs(params.cfg);
-      await sweepIdleRuntimes();
-      if (idleTtlMs > 0) {
-        ensureIdleSweepTimer();
-      }
-      if (params.sessionKey) {
-        sessionIdBySessionKey.set(params.sessionKey, params.sessionId);
-      }
-
-      const fullConfig = loadSessionMcpConfig({
-        workspaceDir: params.workspaceDir,
-        cfg: params.cfg,
-        logDiagnostics: false,
-        manifestRegistry: params.manifestRegistry,
-      });
-      // Safe names from the FULL declared set so partial resolution never changes tool names.
-      const safeServerNamesByServer = assignSafeServerNames(
-        Object.keys(fullConfig.loaded.mcpServers),
-      );
-      const { staticServers, requesterScopedServerNames } = partitionMcpServersByConnectionScope(
-        fullConfig.loaded.mcpServers,
-      );
-      const hasRequesterScoped = requesterScopedServerNames.length > 0;
-
-      if (!hasRequesterScoped) {
-        return await getOrCreateRuntimeEntry({
-          runtimeKey: params.sessionId,
-          sessionId: params.sessionId,
-          sessionKey: params.sessionKey,
-          workspaceDir: params.workspaceDir,
-          agentDir: params.agentDir,
-          cfg: params.cfg,
-          manifestRegistry: params.manifestRegistry,
-          idleTtlMs,
-          safeServerNamesByServer,
-        });
-      }
-
-      const parts: SessionMcpRuntime[] = [];
-      const scopedNameSet = new Set(requesterScopedServerNames);
-      let emptyStaticRuntime: SessionMcpRuntime | undefined;
-      if (Object.keys(staticServers).length > 0) {
-        parts.push(
-          await getOrCreateRuntimeEntry({
-            runtimeKey: params.sessionId,
-            sessionId: params.sessionId,
-            sessionKey: params.sessionKey,
-            workspaceDir: params.workspaceDir,
-            agentDir: params.agentDir,
-            cfg: params.cfg,
-            manifestRegistry: params.manifestRegistry,
-            idleTtlMs,
-            excludeServerNames: scopedNameSet,
-            safeServerNamesByServer,
-          }),
-        );
-      } else {
-        // Reconcile bare key when every server is requester-scoped.
-        emptyStaticRuntime = await getOrCreateRuntimeEntry({
-          runtimeKey: params.sessionId,
-          sessionId: params.sessionId,
-          sessionKey: params.sessionKey,
-          workspaceDir: params.workspaceDir,
-          agentDir: params.agentDir,
-          cfg: params.cfg,
-          manifestRegistry: params.manifestRegistry,
-          idleTtlMs,
-          includeServerNames: new Set(),
-          safeServerNamesByServer,
-        });
-      }
-
-      const requesterSenderId = normalizeOptionalString(params.requesterSenderId);
-      if (requesterSenderId) {
-        const requesterScope: SessionMcpRequesterScope = {
-          requesterSenderId,
-          ...(normalizeOptionalString(params.agentAccountId)
-            ? { agentAccountId: normalizeOptionalString(params.agentAccountId) }
-            : {}),
-          ...(normalizeOptionalString(params.messageChannel)
-            ? { messageChannel: normalizeOptionalString(params.messageChannel) }
-            : {}),
-        };
-        const runtimeKey = buildMcpRequesterRuntimeCacheKey({
-          sessionId: params.sessionId,
-          messageChannel: params.messageChannel,
-          agentAccountId: params.agentAccountId,
-          requesterSenderId,
-        });
-        const { fingerprint: fullScopedFingerprint } = loadSessionMcpConfig({
-          workspaceDir: params.workspaceDir,
-          cfg: params.cfg,
-          logDiagnostics: false,
-          manifestRegistry: params.manifestRegistry,
-          includeServerNames: scopedNameSet,
-          redactConnectionServerNames: scopedNameSet,
-          safeServerNamesByServer,
-        });
-        const scopedRuntime = await runExclusiveOnRuntimeKey(runtimeKey, () =>
-          resolveAndInstallRequesterRuntime({
-            runtimeKey,
-            sessionId: params.sessionId,
-            sessionKey: params.sessionKey,
-            workspaceDir: params.workspaceDir,
-            agentDir: params.agentDir,
-            cfg: params.cfg,
-            manifestRegistry: params.manifestRegistry,
-            idleTtlMs,
-            requesterScopedServerNames,
-            scopedNameSet,
-            safeServerNamesByServer,
-            fullScopedFingerprint,
-            requesterSenderId,
-            agentAccountId: params.agentAccountId,
-            messageChannel: params.messageChannel,
-            requesterScope,
-          }),
-        );
-        if (scopedRuntime) {
-          parts.push(scopedRuntime);
-        }
-        await enforceRequesterRuntimeCap(params.sessionId, runtimeKey);
-      }
-
-      if (parts.length === 0) {
-        return (
-          emptyStaticRuntime ??
-          (await getOrCreateRuntimeEntry({
-            runtimeKey: params.sessionId,
-            sessionId: params.sessionId,
-            sessionKey: params.sessionKey,
-            workspaceDir: params.workspaceDir,
-            agentDir: params.agentDir,
-            cfg: params.cfg,
-            manifestRegistry: params.manifestRegistry,
-            idleTtlMs,
-            includeServerNames: new Set(),
-            safeServerNamesByServer,
-          }))
-        );
-      }
-
-      return createCombinedSessionMcpRuntime({
-        sessionId: params.sessionId,
-        sessionKey: params.sessionKey,
-        workspaceDir: params.workspaceDir,
-        agentDir: params.agentDir,
-        parts,
-      });
-    },
-    async getOrCreateRequesterScoped(params) {
-      // Scoped-only path for shared-thread harnesses: never open static transports
-      // (those stay harness-native) so we do not double-connect.
-      const idleTtlMs = resolveSessionMcpRuntimeIdleTtlMs(params.cfg);
-      await sweepIdleRuntimes();
-      if (idleTtlMs > 0) {
-        ensureIdleSweepTimer();
-      }
-      if (params.sessionKey) {
-        sessionIdBySessionKey.set(params.sessionKey, params.sessionId);
-      }
-      const requesterSenderId = normalizeOptionalString(params.requesterSenderId);
-      if (!requesterSenderId) {
-        return undefined;
-      }
-      const fullConfig = loadSessionMcpConfig({
-        workspaceDir: params.workspaceDir,
-        cfg: params.cfg,
-        logDiagnostics: false,
-        manifestRegistry: params.manifestRegistry,
-      });
-      const { requesterScopedServerNames } = partitionMcpServersByConnectionScope(
-        fullConfig.loaded.mcpServers,
-      );
-      if (requesterScopedServerNames.length === 0) {
-        return undefined;
-      }
-      const safeServerNamesByServer = assignSafeServerNames(
-        Object.keys(fullConfig.loaded.mcpServers),
-      );
-      const scopedNameSet = new Set(requesterScopedServerNames);
-      const requesterScope: SessionMcpRequesterScope = {
-        requesterSenderId,
-        ...(normalizeOptionalString(params.agentAccountId)
-          ? { agentAccountId: normalizeOptionalString(params.agentAccountId) }
-          : {}),
-        ...(normalizeOptionalString(params.messageChannel)
-          ? { messageChannel: normalizeOptionalString(params.messageChannel) }
-          : {}),
-      };
-      const runtimeKey = buildMcpRequesterRuntimeCacheKey({
-        sessionId: params.sessionId,
-        messageChannel: params.messageChannel,
-        agentAccountId: params.agentAccountId,
-        requesterSenderId,
-      });
-      const { fingerprint: fullScopedFingerprint } = loadSessionMcpConfig({
-        workspaceDir: params.workspaceDir,
-        cfg: params.cfg,
-        logDiagnostics: false,
-        manifestRegistry: params.manifestRegistry,
-        includeServerNames: scopedNameSet,
-        redactConnectionServerNames: scopedNameSet,
-        safeServerNamesByServer,
-      });
-      const scopedRuntime = await runExclusiveOnRuntimeKey(runtimeKey, () =>
-        resolveAndInstallRequesterRuntime({
-          runtimeKey,
-          sessionId: params.sessionId,
-          sessionKey: params.sessionKey,
-          workspaceDir: params.workspaceDir,
-          agentDir: params.agentDir,
-          cfg: params.cfg,
-          manifestRegistry: params.manifestRegistry,
-          idleTtlMs,
-          requesterScopedServerNames,
-          scopedNameSet,
-          safeServerNamesByServer,
-          fullScopedFingerprint,
-          requesterSenderId,
-          agentAccountId: params.agentAccountId,
-          messageChannel: params.messageChannel,
-          requesterScope,
-        }),
-      );
-      if (scopedRuntime) {
-        await enforceRequesterRuntimeCap(params.sessionId, runtimeKey);
-      }
-      return scopedRuntime;
-    },
-    rememberAdvertisedScopedCatalog,
-    getAdvertisedScopedCatalog,
-    bindSessionKey(sessionKey, sessionId) {
-      sessionIdBySessionKey.set(sessionKey, sessionId);
-    },
-    resolveSessionId(sessionKey) {
-      return sessionIdBySessionKey.get(sessionKey);
-    },
-    peekSession(params) {
-      const sessionId =
-        params.sessionId ??
-        (params.sessionKey ? sessionIdBySessionKey.get(params.sessionKey) : undefined);
-      return sessionId ? runtimesBySessionId.get(sessionId) : undefined;
-    },
-    async disposeSession(sessionId) {
-      await disposeManagedSession(sessionId);
-    },
-    deferRetirement(sessionId) {
-      if (runtimeKeysForSessionId(sessionId).length === 0) {
-        return false;
-      }
-      deferredRetirementSessionIds.add(sessionId);
-      return true;
-    },
-    async completeDeferredRetirement(sessionId, runtime) {
-      if (!deferredRetirementSessionIds.has(sessionId) || runtime.sessionId !== sessionId) {
-        return false;
-      }
-      if (totalActiveLeasesForSessionId(sessionId) > 0 || (runtime.activeLeases ?? 0) > 0) {
-        return false;
-      }
-      const managed = runtimeKeysForSessionId(sessionId)
-        .map((runtimeKey) => runtimesBySessionId.get(runtimeKey))
-        .filter((entry): entry is SessionMcpRuntime => Boolean(entry));
-      if (managed.length === 0) {
-        return false;
-      }
-      const managedSet = new Set(managed);
-      if (isCombinedSessionMcpRuntime(runtime)) {
-        if (!runtime.managedParts.every((part) => managedSet.has(part))) {
-          return false;
-        }
-      } else if (!managedSet.has(runtime)) {
-        return false;
-      }
-      await disposeManagedSession(sessionId);
-      return true;
-    },
-    async disposeAll() {
-      clearIdleSweepTimer();
-      // Drain all requester chains before clearing maps.
-      const chains = Array.from(requesterWorkChains.values());
-      requesterWorkChains.clear();
-      await Promise.allSettled(chains);
-      const inFlightRuntimes = Array.from(createInFlight.values());
-      createInFlight.clear();
-      const runtimes = Array.from(runtimesBySessionId.values());
-      runtimesBySessionId.clear();
-      sessionIdBySessionKey.clear();
-      idleTtlMsBySessionId.clear();
-      deferredRetirementSessionIds.clear();
-      connectionMetaByRuntimeKey.clear();
-      advertisedScopedCatalogBySessionId.clear();
-      const lateRuntimes = await Promise.all(
-        inFlightRuntimes.map(async ({ promise }) => await promise.catch(() => undefined)),
-      );
-      const allRuntimes = new Set<SessionMcpRuntime>(runtimes);
-      for (const runtime of lateRuntimes) {
-        if (runtime) {
-          allRuntimes.add(runtime);
-        }
-      }
-      await Promise.allSettled(Array.from(allRuntimes, (runtime) => runtime.dispose()));
-    },
-    sweepIdleRuntimes,
-    listSessionIds() {
-      return [
-        ...new Set(Array.from(runtimesBySessionId.values(), (runtime) => runtime.sessionId)),
-      ].toSorted((a, b) => a.localeCompare(b));
-    },
-    listRuntimeKeys() {
-      return Array.from(runtimesBySessionId.keys()).toSorted((a, b) => a.localeCompare(b));
-    },
-    totalActiveLeasesForSession(sessionId) {
-      return totalActiveLeasesForSessionId(sessionId);
-    },
-  };
-  // Test-only bookkeeping snapshot for drain assertions.
-  Object.assign(manager, {
-    bookkeepingSizesForTest: () => ({
-      runtimes: runtimesBySessionId.size,
-      connectionMeta: connectionMetaByRuntimeKey.size,
-      createInFlight: createInFlight.size,
-      requesterWorkChains: requesterWorkChains.size,
-      sessionKeys: sessionIdBySessionKey.size,
-      idleTtl: idleTtlMsBySessionId.size,
-      deferredRetirement: deferredRetirementSessionIds.size,
-      advertisedScopedCatalogs: advertisedScopedCatalogBySessionId.size,
-    }),
-  });
-  return manager;
-}
-
-function getSessionMcpRuntimeManager(): SessionMcpRuntimeManager {
-  return resolveGlobalSingleton(SESSION_MCP_RUNTIME_MANAGER_KEY, createSessionMcpRuntimeManager);
-}
-
-export async function getOrCreateSessionMcpRuntime(params: {
-  sessionId: string;
-  sessionKey?: string;
-  workspaceDir: string;
-  agentDir?: string;
-  cfg?: OpenClawConfig;
-  manifestRegistry?: Pick<PluginManifestRegistry, "plugins">;
-  requesterSenderId?: string | null;
-  agentAccountId?: string | null;
-  messageChannel?: string | null;
-}): Promise<SessionMcpRuntime> {
-  return await getSessionMcpRuntimeManager().getOrCreate(params);
-}
-
-/**
- * Requester-scoped MCP runtime only (no static partition).
- * Shared-thread harnesses use this so static MCP stays harness-native.
- */
-export async function getOrCreateRequesterScopedMcpRuntime(params: {
-  sessionId: string;
-  sessionKey?: string;
-  workspaceDir: string;
-  agentDir?: string;
-  cfg?: OpenClawConfig;
-  manifestRegistry?: Pick<PluginManifestRegistry, "plugins">;
-  requesterSenderId?: string | null;
-  agentAccountId?: string | null;
-  messageChannel?: string | null;
-}): Promise<SessionMcpRuntime | undefined> {
-  return await getSessionMcpRuntimeManager().getOrCreateRequesterScoped(params);
-}
-
-export function rememberAdvertisedScopedMcpCatalog(
-  sessionId: string,
-  catalog: McpToolCatalog,
-): void {
-  getSessionMcpRuntimeManager().rememberAdvertisedScopedCatalog(sessionId, catalog);
-}
-
-export function getAdvertisedScopedMcpCatalog(sessionId: string): McpToolCatalog | null {
-  return getSessionMcpRuntimeManager().getAdvertisedScopedCatalog(sessionId);
-}
-
-/** Looks up an existing session MCP runtime without creating it or connecting transports. */
-export function peekSessionMcpRuntime(params: {
-  sessionId?: string | null;
-  sessionKey?: string | null;
-}): SessionMcpRuntime | undefined {
-  const sessionId = normalizeOptionalString(params.sessionId);
-  const sessionKey = normalizeOptionalString(params.sessionKey);
-  return getSessionMcpRuntimeManager().peekSession({
-    ...(sessionId ? { sessionId } : {}),
-    ...(sessionKey ? { sessionKey } : {}),
-  });
-}
-
-async function disposeSessionMcpRuntime(sessionId: string): Promise<void> {
-  await getSessionMcpRuntimeManager().disposeSession(sessionId);
-}
-
-export async function retireSessionMcpRuntime(params: {
-  sessionId?: string | null;
-  reason: string;
-  preserveActiveLeases?: boolean;
-  onError?: (error: unknown, sessionId: string, reason: string) => void;
-}): Promise<boolean> {
-  const sessionId = normalizeOptionalString(params.sessionId);
-  if (!sessionId) {
-    return false;
-  }
-  const manager = getSessionMcpRuntimeManager();
-  // Aggregate leases across static + all requester-scoped parts so preserveActiveLeases
-  // does not miss a leased scoped runtime while peeking only the bare session key.
-  if (params.preserveActiveLeases === true && manager.totalActiveLeasesForSession(sessionId) > 0) {
-    manager.deferRetirement(sessionId);
-    return true;
-  }
-  try {
-    await disposeSessionMcpRuntime(sessionId);
-    return true;
-  } catch (error) {
-    params.onError?.(error, sessionId, params.reason);
-    return false;
-  }
-}
-
-/** Completes a one-shot retirement after its final run, view, or request lease releases. */
-export async function completeDeferredSessionMcpRuntimeRetirement(
-  runtime: SessionMcpRuntime,
-): Promise<boolean> {
-  return await getSessionMcpRuntimeManager().completeDeferredRetirement(runtime.sessionId, runtime);
-}
-
-export async function retireSessionMcpRuntimeForSessionKey(params: {
-  sessionKey?: string | null;
-  reason: string;
-  preserveActiveLeases?: boolean;
-  onError?: (error: unknown, sessionId: string, reason: string) => void;
-}): Promise<boolean> {
-  const sessionKey = normalizeOptionalString(params.sessionKey);
-  if (!sessionKey) {
-    return false;
-  }
-  const sessionId = getSessionMcpRuntimeManager().resolveSessionId(sessionKey);
-  return await retireSessionMcpRuntime({
-    sessionId,
-    reason: params.reason,
-    preserveActiveLeases: params.preserveActiveLeases,
-    onError: params.onError,
-  });
-}
-
-export async function disposeAllSessionMcpRuntimes(): Promise<void> {
-  await getSessionMcpRuntimeManager().disposeAll();
-}
+export { createSessionMcpRuntimeManager };
+export { mergeMcpToolCatalogs };
 
 export const testing = {
   buildMcpClientCapabilities,
@@ -2384,10 +978,10 @@ export const testing = {
     resolverTesting.setMcpConnectionRevalidateMsForTest();
   },
   getCachedSessionIds() {
-    return getSessionMcpRuntimeManager().listSessionIds();
+    return getSessionMcpRuntimeManagerForTesting().listSessionIds();
   },
   getCachedRuntimeKeys() {
-    return getSessionMcpRuntimeManager().listRuntimeKeys();
+    return getSessionMcpRuntimeManagerForTesting().listRuntimeKeys();
   },
   getBookkeepingSizes(manager: SessionMcpRuntimeManager): Record<string, number> {
     const sizes = (
