+++ src/agents/agent-bundle-mcp-tools.ts
@@ -2,11 +2,8 @@
 export type { McpToolCatalogDiagnostic } from "./agent-bundle-mcp-types.js";
 export {
   disposeAllSessionMcpRuntimes,
-  getAdvertisedScopedMcpCatalog,
-  getOrCreateRequesterScopedMcpRuntime,
   getOrCreateSessionMcpRuntime,
   peekSessionMcpRuntime,
-  rememberAdvertisedScopedMcpCatalog,
   resolveSessionMcpConfigSummary,
   retireSessionMcpRuntime,
   retireSessionMcpRuntimeForSessionKey,
@@ -16,8 +13,3 @@ export {
   createBundleMcpToolRuntime,
   materializeBundleMcpToolsForRun,
 } from "./agent-bundle-mcp-materialize.js";
-export {
-  materializeRequesterScopedMcpToolsForHarnessRun,
-  type MaterializeRequesterScopedMcpToolsForHarnessRunParams,
-  type RequesterScopedHarnessMcpTools,
-} from "./agent-bundle-mcp-harness.js";
